THE IMPACT OF MATHEMATICS ON PROGRAMMING
Programming is the art of telling the computer what to do using some set of laid down rules and principles to make digital applications e.g Softwares, Apps, Websites, and Operating systems. The argument of whether Programming is an Art or Science will be a topic to be discussed another day but for now, let's just call it an Art since Programmers write code.