REST API comprehensive Guide.
The article is about rest API's and how they work. It consists of various code examples as well. It starts from the topic What is rest Api and Why we need that. The article tells us the underlying principles of Rest API as well and there is also a link to referenced status code.