Renamed hero.png back to image1.png
I did not intend to have image1 be the hero image.

I'm am not sure if you now want me to include the hero image as part of my commit so I made one that should work.
I am not super happy with it as I just took an image of my laptop with my phone. So, if you want to find a better one fell free to do so.