Added index.md file and hero.jpg
This article describes RPA. It explains how it differs from traditional automation in terms of features and functionality. It also explains the importance of RPA and why we should consider it for enterprises. The main take away from the article is how RPA differs from automation and how it makes business processes easier and cheaper.